+++ ChangeLog.rst
@@ -3,7 +3,7 @@ development
 
  * deprecated no_atexit in favor of atexit keyword parameter, 
    https://github.com/amitu/importd/issues/1
- 
+ * gunicorn cleanly exits now
 
 0.1.4 - 22-Oct-2012
 ===================